backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.38.35",
    "flask>=3.1.1",
    "flask-cors>=6.0.1",
    "python-dotenv>=1.1.0",
]
